+++ src/main/java/org/folio/inventory/domain/Contributor.java
@@ -1,0 +1,11 @@
+package org.folio.inventory.domain;
+
+public class Contributor {
+  public final String contributorNameTypeId;
+  public final String name;
+
+  public Contributor(String contributorNameTypeId, String name) {
+    this.contributorNameTypeId = contributorNameTypeId;
+    this.name = name;
+  }
+}