+++ src/main/java/org/folio/inventory/resources/Items.java
@@ -222,7 +222,7 @@ public class Items {
               String permanentLocationId = holdingResponse.getStatusCode() == 200
                 && holdingResponse.getJson().containsKey("permanentLocationId")
                 ? holdingResponse.getJson().getString("permanentLocationId")
-                : item.permanentLocationId;
+                : null;
 
               ArrayList<CompletableFuture<Response>> allFutures = new ArrayList<>();
 
@@ -273,25 +273,21 @@ public class Items {
     List<String> notes;
     notes = JsonArrayHelper.toListOfStrings(itemRequest.getJsonArray("notes"));
     String materialTypeId = getNestedProperty(itemRequest, "materialType", "id");
-    String permanentLocationId = getNestedProperty(itemRequest, "permanentLocation", "id");
     String temporaryLocationId = getNestedProperty(itemRequest, "temporaryLocation", "id");
     String permanentLoanTypeId = getNestedProperty(itemRequest, "permanentLoanType", "id");
     String temporaryLoanTypeId = getNestedProperty(itemRequest, "temporaryLoanType", "id");
 
     return new Item(
       itemRequest.getString("id"),
-      itemRequest.getString("title"),
       itemRequest.getString("barcode"),
       itemRequest.getString("enumeration"),
       itemRequest.getString("chronology"),
       pieceIdentifiers,
       itemRequest.getString("numberOfPieces"),
-      itemRequest.getString("instanceId"),
       itemRequest.getString("holdingsRecordId"),
       notes,
       status,
       materialTypeId,
-      permanentLocationId,
       temporaryLocationId,
       permanentLoanTypeId,
       temporaryLoanTypeId);
@@ -426,7 +422,7 @@ public class Items {
           });
 
         List<String> permanentLocationIds = wrappedItems.records.stream()
-          .map(item -> HoldingsSupport.determinePermanentLocationIdForItem(item,
+          .map(item -> HoldingsSupport.determinePermanentLocationIdForItem(
             HoldingsSupport.holdingForItem(item, holdings).orElse(null)))
           .filter(Objects::nonNull)
           .distinct()